Moved prometheus related function to separate go file
